Vif: Fix for Unpacks when WL = 0. (KH2 + Tests) Removes a long standing hack.
Note: CL = 0 behaviour is still not completely accurate, the first vector is incorrect, but will look at that another day, need a game that does it first really so we can see if it helps :)

+++ pcsx2/x86/newVif.h
@@ -61,9 +61,9 @@ struct __aligned16 nVifBlock {
 	u8   num;		// [00] Num  Field
 	u8   upkType;	// [01] Unpack Type [usn*1:mask*1:upk*4]
 	u8   mode;		// [02] Mode Field
-	u8   cl;		// [03] CL   Field
+	u16   cl;		// [03] CL   Field
 	u32  mask;		// [04] Mask Field
-	u8   wl;		// [08] WL   Field
+	u16   wl;		// [08] WL   Field
 	u8   aligned;   // [09] Packet Alignment
 	u8	 padding[2];// [10] through [11]
 	uptr startPtr;	// [12] Start Ptr of RecGen Code
